feat: create homeviewmodel, create groupletter section subview, adjust home view
